# contracts/agent-runtime/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Agent Runtime HTTP API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /team-runs:
    post:
      operationId: startAgentTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartAgentTeamRequest'
      responses:
        '202':
          description: Accepted coordinator-led Agent Team with durable member handoffs and fan-in join
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentTeamStartResult'
        default:
          $ref: '#/components/responses/Error'
...
components:
...
  responses:
    Error:
      description: Stable Runtime error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ActorRef:
      type: object
      additionalProperties: false
      required: [id]
      properties:
        tenantID: {type: string}
        id: {type: string}
    ThreadRef:
      type: object
      additionalProperties: false
      required: [kind, id]
      properties:
        kind: {type: string}
        id: {type: string}
    ProjectionRef:
      type: object
      additionalProperties: false
      required: [kind, id]
      properties:
        kind: {type: string}
        id: {type: string}
    ResourceRef:
      type: object
      additionalProperties: false
      required: [kind, id]
      properties:
        kind: {type: string}
        id: {type: string}
        revision: {type: string}
...
    RunThread:
      type: object
      additionalProperties: false
      required: [kind, id]
      properties:
        kind: {type: string}
        id: {type: string}
        parentProjection:
          $ref: '#/components/schemas/ProjectionRef'
        sourceProjection:
          $ref: '#/components/schemas/ProjectionRef'
        branchReason:
          type: string
          enum: [default, retry, edit]
    RunInput:
      type: object
      additionalProperties: false
      required: [content, contentType]
      properties:
        content: {type: string, maxLength: 20000}
        contentType: {type: string, enum: [text, markdown]}
        fileIDs: {type: array, maxItems: 20, items: {type: string}}
        outputIDs: {type: array, maxItems: 50, items: {type: string}}
        evidenceIDs: {type: array, maxItems: 50, items: {type: string}}
        htmlVisualPrompt: {type: boolean}
        htmlVisualColorMode: {type: string, enum: [light, dark]}
...
    WorkspaceExtension:
      type: object
      required: [schemaVersion, type]
      additionalProperties: true
      properties:
        schemaVersion: {type: integer, minimum: 1}
        type: {type: string}
        resourceID:
          type: string
          maxLength: 128
          description: Optional host-authorized resource override, used for isolated Story branches.
...
    AgentTeamMemberRequest:
      type: object
      additionalProperties: false
      required: [memberID, agentManifest, goal]
      properties:
        memberID: {type: string, minLength: 1, maxLength: 64}
        agentManifest:
          $ref: '#/components/schemas/ResourceRef'
        goal: {type: string, minLength: 1, maxLength: 20000}
        contentType: {type: string, enum: [text, markdown]}
        outputIDs: {type: array, maxItems: 128, items: {type: string}}
        evidenceIDs: {type: array, maxItems: 128, items: {type: string}}
        options: {type: object, additionalProperties: true}
        maxLLMCalls: {type: integer, minimum: 1, maximum: 32}
        maxToolCalls: {type: integer, minimum: 1, maximum: 64}
        structuredOutputSchema: {type: object, additionalProperties: true}
        resultAttempts: {type: integer, minimum: 1, maximum: 2}
    AgentTeamJoinRequest:
      type: object
      additionalProperties: false
      properties:
        mode: {type: string, enum: [all, any, quorum], default: all}
        quorum: {type: integer, minimum: 1, maximum: 16}
        failurePolicy: {type: string, enum: [collect, fail_fast], default: collect}
        timeoutSeconds: {type: integer, minimum: 60, maximum: 604800, default: 3600}
        timeoutPolicy: {type: string, enum: [cancel_pending, leave_running], default: cancel_pending}
    StartAgentTeamRequest:
      type: object
      additionalProperties: false
      required: [thread, input, clientTeamID, coordinatorManifest, members, join]
      properties:
        thread:
          $ref: '#/components/schemas/RunThread'
        input:
          $ref: '#/components/schemas/RunInput'
        clientTeamID: {type: string, minLength: 1, maxLength: 128}
        coordinatorManifest:
          $ref: '#/components/schemas/ResourceRef'
        model: {type: string, maxLength: 128}
        options: {type: object, additionalProperties: true}
        workspace:
          $ref: '#/components/schemas/WorkspaceExtension'
        maxLLMCalls: {type: integer, minimum: 1, maximum: 32}
        maxToolCalls: {type: integer, minimum: 1, maximum: 64}
        structuredOutputSchema: {type: object, additionalProperties: true}
        resultAttempts: {type: integer, minimum: 1, maximum: 2}
        members:
          type: array
          minItems: 1
          maxItems: 16
          items:
            $ref: '#/components/schemas/AgentTeamMemberRequest'
        join:
          $ref: '#/components/schemas/AgentTeamJoinRequest'
...
    Run:
      type: object
      additionalProperties: true
      required: [schemaVersion, runtimeKind, actor, thread, runID, requestID, status]
      properties:
        schemaVersion: {type: integer, enum: [1]}
        runtimeKind: {type: string, enum: [text, plan_execute, workflow, team]}
        actor:
          $ref: '#/components/schemas/ActorRef'
        thread:
          $ref: '#/components/schemas/ThreadRef'
        runID: {type: string}
        requestID: {type: string}
        status: {type: string}
        agentManifestRef:
          $ref: '#/components/schemas/ResourceRef'
        workflowDefinitionRef:
          $ref: '#/components/schemas/ResourceRef'
        agentName: {type: string}
        rootRunID: {type: string}
        parentRunID: {type: string}
        handoffID: {type: string}
        depth: {type: integer, minimum: 1}
...
    RunStep:
      type: object
      additionalProperties: true
      required: [stepID, runID, status]
      properties:
        stepID: {type: string}
        runID: {type: string}
        status: {type: string}
        nodeID: {type: string}
        activationPath: {type: string}
        laneID: {type: string}
        activationIndex: {type: integer, minimum: 0}
        completionOrder: {type: integer, minimum: 0}
        waitKind: {type: string}
        waitID: {type: string}
        childRunID: {type: string}
...
    AgentTeamTaskStart:
      type: object
      additionalProperties: false
      required: [memberID, handoff, run, rootStep]
      properties:
        memberID: {type: string}
        handoff:
          $ref: '#/components/schemas/RunHandoff'
        run:
          $ref: '#/components/schemas/Run'
        rootStep:
          $ref: '#/components/schemas/RunStep'
    AgentTeamStartResult:
      type: object
      additionalProperties: false
      required: [rootRun, rootStep, inputProjectionRef, outputProjectionRef, tasks, join]
      properties:
        rootRun:
          $ref: '#/components/schemas/Run'
        rootStep:
          $ref: '#/components/schemas/RunStep'
        inputProjectionRef:
          $ref: '#/components/schemas/ProjectionRef'
        outputProjectionRef:
          $ref: '#/components/schemas/ProjectionRef'
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/AgentTeamTaskStart'
        join:
          $ref: '#/components/schemas/RunHandoffJoin'
    RunHandoff:
      type: object
      additionalProperties: false
      required: [handoffID, clientHandoffID, rootRunID, parentRunID, childRunID, agentManifest, agentName, goal, status, depth, inputProjectionRef, resultOutputIDs, createdAt, updatedAt]
      properties:
        handoffID: {type: string}
        clientHandoffID: {type: string}
        rootRunID: {type: string}
        parentRunID: {type: string}
        childRunID: {type: string}
        agentManifest:
          $ref: '#/components/schemas/ResourceRef'
        agentName: {type: string}
        goal: {type: string}
        status: {type: string, enum: [queued, completed, failed, cancelled]}
        depth: {type: integer, minimum: 1}
        inputProjectionRef:
          $ref: '#/components/schemas/ProjectionRef'
        resultSummary: {type: string}
        resultOutputIDs: {type: array, items: {type: string}}
        errorCode: {type: string}
        errorMessage: {type: string}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        completedAt: {type: string, format: date-time, nullable: true}
    RunHandoffJoin:
      type: object
      additionalProperties: false
      required: [joinID, clientJoinID, rootRunID, parentRunID, handoffIDs, mode, quorum, failurePolicy, timeoutSeconds, timeoutPolicy, deadlineAt, status, completedCount, failedCount, cancelledCount, pendingCount, resultHandoffIDs, createdAt, updatedAt]
      properties:
        joinID: {type: string}
        clientJoinID: {type: string}
        rootRunID: {type: string}
        parentRunID: {type: string}
        handoffIDs:
          type: array
          items: {type: string}
        mode: {type: string, enum: [all, any, quorum]}
        quorum: {type: integer, minimum: 1}
        failurePolicy: {type: string, enum: [collect, fail_fast]}
        timeoutSeconds: {type: integer, minimum: 60, maximum: 604800}
        timeoutPolicy: {type: string, enum: [cancel_pending, leave_running]}
        deadlineAt: {type: string, format: date-time}
        status: {type: string, enum: [pending, ready, failed, cancelled]}
        completedCount: {type: integer, minimum: 0}
        failedCount: {type: integer, minimum: 0}
        cancelledCount: {type: integer, minimum: 0}
        pendingCount: {type: integer, minimum: 0}
        resultHandoffIDs:
          type: array
          items: {type: string}
        errorCode: {type: string}
        errorMessage: {type: string}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        resolvedAt: {type: string, format: date-time, nullable: true}
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestID]
          properties:
            code: {type: string}
            message: {type: string}
            requestID: {type: string}